% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIPD.R
\name{getIPD}
\alias{getIPD}
\title{Reconstruct individual patient data (IPD) from Kaplan-Meier (KM) survival curves}
\usage{
getIPD(n, t, S, cens.t = NA)
}
\arguments{
\item{n}{The number at risk at time zero.}

\item{t}{A vector of the x coordinates of the bottoms of the drops in survival.}

\item{S}{A vector of the y coordinates corresponding to 't', between 0 and 1. These represent survival proportions at the times in t.}

\item{cens.t}{A vector of times at which patients were censored (optional).}
}
\value{
A data.frame of reconstructed IPD with two columns; (1) t: survival time, (2) event: indicator of event, where 1=event and 0=censored.
}
\description{
This function returns reconstructed individual patient data (IPD) based on digitized data from Kaplan-Meier (KM) survival curves. This function takes as an input a vector of digitized censoring times and will incorporate them into the IPD exactly as specified.
}
\details{
If the KM curve is horizontal at its end, either (1) the coordinates of the end of the curve can be included in t and S, OR (2) the x-coordinate of the end of the curve can be included in cens.t.
}
